src/requirements.txt
wavelink
Levenshtein~=0.20.8
boto3~=1.26.8
discord.py~=2.1.0
google-api-python-client~=2.65.0
PyNaCl~=1.5.0
pymongo~=4.3.3
pyyaml~=6.0
riotwatcher~=3.2.4
jira
flask
requests
jira